refactor(cedar-hitl): dedup formatMinuteBucket + tighten GetPendingFn IAM
S5 (formatMinuteBucket duplicated in 5 handlers) — extracted to
``cdk/src/handlers/shared/rate-limit.ts``. The same identical
``yyyymmddhhmm`` UTC-bucket helper was inlined in approve-task,
deny-task, get-pending, get-policies, and nudge-task — divergence
here would silently invalidate per-minute rate-limit counters
across handlers. Centralising it pins the bucket boundary.

S6 (GetPendingFn over-permissioned) — replaced
``grantReadWriteData`` with two explicit ``PolicyStatement`` grants:
``dynamodb:Query`` on the table + GSI ARNs (the only read pattern
the handler uses), and ``dynamodb:UpdateItem`` on just the table
ARN (for the synthetic ``RATE#<user_id>#PENDING`` rate-limit row).
Drops PutItem, BatchWrite, DeleteItem, and BatchGet permissions
that were latent in the prior wildcard grant — none of which the
handler ever exercises. Pinned to specific ARNs (not ``/*``).

Tests: jest on the 5 touched handlers + task-api construct — 116
passed across 6 suites.
